requirements.txt
regex
requests
llama_parse
google.generativeai
langchain_google_genai
